LAB 3,4
Principles of object-oriented programming SOLID and STUPID.
The class of the object. Implementing its default methods.
Features of the implementation of inheritance in Java. Simple and multiple inheritance.
The concept of an abstract class. Modifier.
The concept of the interface. Implementation of interfaces in Java, default methods. Differences from abstract classes.
Enumerated data type (enum) in Java. Features of implementation and use.
Methods and fields with modifiers are static and final.
Overloading and redefining methods. Covariates of the returned data types.
Elements of functional programming in the Java syntax. Functional interfaces, lambda expressions. References to methods.
Exception handling, three types of exceptions.
Nested, local, and anonymous classes.
The mechanism of reflection in Java. Class Class.